Extract DetailStack tests
